+++ smc/vpn/route.py
@@ -3,67 +3,127 @@
     Route based VPNs with multi-domain support, requires SMC >=6.3
 
 Module for configuring Route Based VPN.
-When creating RBVPN local and remote endpoints, use the TunnelEndpoint
-create classmethods based on GRE or IPSEC VPN. 
+Creating a route based VPN consists of creating a local and remote tunnel
+endpoint. Once you have the required endpoints, use TunnelEndpoint
+classmethods to create the VPN by type (i.e. GRE, IPSEC).
 
 List all existing route based VPNs::
 
     print(list(RouteVPN.objects.all()))
     
-Example of creating an IPSEC wrapped RBVPN using a third party remote GW::
+Example of fully provisioning an IPSEC wrapped RBVPN using a third
+party remote GW::
 
-    engine = Engine('fw1')
+    engine = Layer3Firewall.create(
+        name='fw1', mgmt_ip='1.1.1.1', mgmt_network='1.1.1.0/24')
     
-    # Internal Gateway and Tunnel Interface
-    gw = engine.internal_gateway
+    # RBVPN requires a tunnel interface
+    engine.tunnel_interface.add_single_node_interface(
+        tunnel_id=1000,
+        address='2.2.2.2',
+        network_value='2.2.2.0/24')
+    
+    # Enable the internal VPN gateway
+    for vpn in engine.vpn_endpoint:
+        vpn.update(enabled=True)
+        
+    # Create the local Tunnel Endpoint using the engine internal gateway
+    # and previously created tunnel interface
     tunnel_if = engine.tunnel_interface.get(1000)
-    local_gateway = TunnelEndpoint.create_ipsec_endpoint(gw, tunnel_if)
+    local_gateway = TunnelEndpoint.create_ipsec_endpoint(engine.internal_gateway, tunnel_if)
     
-    # Specify the pre-created 3rd party external gateway
-    external = ExternalGateway('mygw')
-    remote_gateway = TunnelEndpoint.create_ipsec_endpoint(external)
+    # Create the remote side network elements
+    Network.create(name='remotenet', ipv4_network='172.18.10.0/24')
+    
+    # An ExternalGateway defines the remote side details
+    # Create the element, define the remote IP and network on other end
+    gw = ExternalGateway.create(name='remotegw')
+    gw.external_endpoint.create(name='endpoint1', address='10.10.10.10')
+    gw.vpn_site.create(name='remotesite', site_element=[Network('remotenet')])
+    
+    # Create the remote Tunnel Endpoint using the external gateway
+    remote_gateway = TunnelEndpoint.create_ipsec_endpoint(gw)
     
     RouteVPN.create_ipsec_tunnel(
-        name='avpn', 
+        name='myvpn', 
         local_endpoint=local_gateway, 
         remote_endpoint=remote_gateway)
+        
 
 Create a GRE Tunnel Mode RBVPN with a remote gateway (non-SMC managed)::
 
-    engine = Engine('foo')
+    engine = Engine('fw') 
     
-    # Enable VPN endpoint on interface 0
-    # Note: An interface can have multiple IP addresses in which case you
-    # may want to get the VPN endpoint match by address
-    vpn_endpoint = None
-    for endpoint in list(engine.vpn_endpoint):
-        if endpoint.physical_interface.interface_id == '0':
-            endpoint.enabled = True
-            endpoint.update()
+    # Enable VPN endpoint on interface 0 
+    # Note: An interface can have multiple IP addresses in which case you 
+    # may want to get the VPN endpoint match by address 
+    vpn_endpoint = None 
+    for endpoint in engine.vpn_endpoint: 
+        if endpoint.physical_interface.interface_id == '0': 
+            endpoint.update(enabled=True) 
             vpn_endpoint = endpoint
+            break
+     
+    # Create a new Tunnel Interface for the engine 
+    engine.tunnel_interface.add_single_node_interface( 
+        tunnel_id=3000, address='30.30.30.30', network_value='30.30.30.0/24') 
     
-    # Create a new Tunnel Interface for the engine
-    engine.tunnel_interface.add_single_node_interface(
-        tunnel_id=3000, address='30.30.30.30', network_value='30.30.30.0/24')
+    tunnel_interface =  engine.tunnel_interface.get(3000)
+    local_endpoint = TunnelEndpoint.create_gre_tunnel_endpoint( 
+        endpoint=vpn_endpoint, tunnel_interface=tunnel_interface) 
+ 
+    # Create GRE tunnel endpoint for remote gateway 
+    remote_endpoint = TunnelEndpoint.create_gre_tunnel_endpoint( 
+        remote_address='10.1.1.2') 
     
-    local_endpoint = TunnelEndpoint.create_gre_tunnel_endpoint(
-        endpoint=vpn_endpoint, tunnel_interface=engine.tunnel_interface.get(3000))
+    # Create the top level IPSEC tunnel to encapsulate RBVPN 
+    policy_vpn = PolicyVPN.create(name='myIPSEC') 
+     
+    RouteVPN.create_gre_tunnel_mode( 
+        name='mytunnelvpn',  
+        local_endpoint=local_endpoint,  
+        remote_endpoint=remote_endpoint,  
+        policy_vpn=policy_vpn) 
 
-    # Create GRE tunnel endpoint for remote gateway
-    remote_endpoint = TunnelEndpoint.create_gre_tunnel_endpoint(
-        remote_address='1.1.1.1')
+Create a no-encryption GRE route based VPN between two managed NGFWs::
+
+    engine1 = Layer3Firewall.create(name='engine1', mgmt_ip='1.1.1.1', mgmt_network='1.1.1.0/24')
+    engine1.tunnel_interface.add_single_node_interface( 
+        tunnel_id=1000, 
+        address='2.2.2.2', 
+        network_value='2.2.2.0/24')
     
-    policy_vpn = PolicyVPN.create(name='myIPSEC')
-    
-    RouteVPN.create_gre_tunnel(
-        name='mytunnelvpn', 
-        local_endpoint=local_endpoint, 
-        remote_endpoint=remote_endpoint, 
-        policy_vpn=policy_vpn)
+    # Obtain the 'internal endpoint' from the NGFW and enable VPN
+    for vpn in engine1.vpn_endpoint:
+        internal_endpoint = vpn
+        vpn.update(enabled=True)
         
+    tunnel_if = engine1.tunnel_interface.get(1000) 
+    local_gateway = TunnelEndpoint.create_gre_tunnel_endpoint(
+        internal_endpoint, tunnel_if)
+    
+    engine2 = Layer3Firewall.create(name='engine2', mgmt_ip='1.1.1.1', mgmt_network='1.1.1.0/24')
+    engine2.tunnel_interface.add_single_node_interface( 
+        tunnel_id=1000, 
+        address='2.2.2.2', 
+        network_value='2.2.2.0/24')
+    
+    # Obtain the 'internal endpoint' from the NGFW and enable VPN
+    for vpn in engine2.vpn_endpoint:
+        internal_endpoint = vpn
+        vpn.update(enabled=True)
+    
+    tunnel_if = engine2.tunnel_interface.get(1000) 
+    remote_gateway = TunnelEndpoint.create_gre_tunnel_endpoint(
+        internal_endpoint, tunnel_if)
+    
+    RouteVPN.create_gre_tunnel_no_encryption(
+        name='openvpn', 
+        local_endpoint=local_gateway,
+        remote_endpoint=remote_gateway)
+
 """
 from smc.base.model import Element, ElementCreator
-from smc.core.interfaces import TunnelInterface
 from smc.core.engine import InternalEndpoint
 from smc.vpn.elements import VPNProfile
 from smc.api.exceptions import CreateElementFailed, CreateVPNFailed
@@ -77,9 +137,9 @@ class RouteVPN(Element):
     
     @classmethod
     def create_ipsec_tunnel(cls, name, local_endpoint, remote_endpoint,
-                            monitoring_group=None, vpn_profile=None,
-                            mtu=0, pmtu_discovery=True, ttl=0,
-                            comment=None):
+                            preshared_key, monitoring_group=None,
+                            vpn_profile=None, mtu=0, pmtu_discovery=True,
+                            ttl=0, enabled=True, comment=None):
         """
         The VPN tunnel type negotiates IPsec tunnels in the same way
         as policy-based VPNs, but traffic is selected to be sent into
@@ -97,6 +157,7 @@ class RouteVPN(Element):
         :param int mtu: Set MTU for this VPN tunnel (default: 0)
         :param boolean pmtu_discovery: enable pmtu discovery (default: True)
         :param int ttl: ttl for connections on the VPN (default: 0)
+        :param bool enabled: enable the RBVPN or leave it disabled
         :param str comment: optional comment
         :raises CreateVPNFailed: failed to create the VPN with reason
         :return RouteVPN
@@ -109,14 +170,16 @@ class RouteVPN(Element):
             'name': name,
             'mtu': mtu,
             'ttl': ttl,
+            'enabled': enabled,
             'monitoring_group_ref': group.href,
             'pmtu_discovery': pmtu_discovery,
+            'preshared_key': preshared_key,
             'rbvpn_tunnel_side_a': local_endpoint.data,
             'rbvpn_tunnel_side_b': remote_endpoint.data,
             'tunnel_mode': 'vpn',
-            'vpn_profile_ref': profile.href,
-            'virtual_interface_vpn_ref': 'http://172.18.1.154:8082/6.3/elements/vpn/2'
-            }
+            'comment': comment,
+            'vpn_profile_ref': profile.href
+        }
         
         try:
             return ElementCreator(cls, json)
@@ -124,12 +187,14 @@ class RouteVPN(Element):
             raise CreateVPNFailed(err)
         
     @classmethod
-    def create_gre_tunnel(cls, name, local_endpoint, remote_endpoint,
-                          policy_vpn, monitoring_group=None, mtu=0,
-                          pmtu_discovery=True, ttl=0, preshared_key=None,
-                          comment=None):
+    def create_gre_tunnel_mode(cls, name, local_endpoint, remote_endpoint,
+                          policy_vpn, mtu=0, pmtu_discovery=True, ttl=0,
+                          enabled=True, comment=None):
         """
-        Create a GRE based route VPN.
+        Create a GRE based tunnel mode route VPN. Tunnel mode GRE wraps the
+        GRE tunnel in an IPSEC tunnel to provide encrypted end-to-end
+        security. Therefore a policy based VPN is required to 'wrap' the
+        GRE into IPSEC. 
         
         :param str name: name of VPN
         :param TunnelEndpoint local_endpoint: the local side endpoint for
@@ -146,35 +211,46 @@ class RouteVPN(Element):
         :raises CreateVPNFailed: failed to create the VPN with reason
         :return RouteVPN
         """
-        group = monitoring_group if monitoring_group else \
-            TunnelMonitoringGroup('Uncategorized')
-        
         json = {
             'name': name,
             'ttl': ttl,
             'mtu': mtu,
             'pmtu_discovery': pmtu_discovery,
-            'preshared_key': '',
             'tunnel_encryption': 'tunnel_mode',
             'tunnel_mode': 'gre',
-            'tunnel_mode_vpn_ref': policy_vpn.href,
-            'monitoring_group_ref': group.href,
+            'enabled': enabled,
+            'comment': comment,
             'rbvpn_tunnel_side_a': local_endpoint.data,
             'rbvpn_tunnel_side_b': remote_endpoint.data
-            }
-        
-        from pprint import pprint
-        pprint(json)
+        }
+        if policy_vpn is None:
+            json['tunnel_encryption'] = 'no_encryption'
+        else:
+            json['tunnel_mode_vpn_ref'] = policy_vpn.href
+          
         try:
             return ElementCreator(cls, json)
         except CreateElementFailed as err:
             raise CreateVPNFailed(err)
-            
+    
     @classmethod
-    def create_gre_transport(cls, name, local_endpoint, remote_endpoint,
+    def create_gre_tunnel_no_encryption(cls, name, local_endpoint, remote_endpoint,
+                          mtu=0, pmtu_discovery=True, ttl=0,
+                          enabled=True, comment=None):
+        """
+        Create a GRE Tunnel with no encryption. See `create_gre_tunnel_mode` for
+        constructor descriptions.
+        """
+        return cls.create_gre_tunnel_mode(
+            name, local_endpoint, remote_endpoint, policy_vpn=None,
+            mtu=mtu, pmtu_discovery=pmtu_discovery, ttl=ttl,
+            enabled=enabled, comment=comment)
+               
+    @classmethod
+    def create_gre_transport_mode(cls, name, local_endpoint, remote_endpoint,
                              preshared_key, monitoring_group=None,
                              vpn_profile=None, mtu=0, ttl=0,
-                             pmtu_discovery=True, comment=None):
+                             pmtu_discovery=True, enabled=True, comment=None):
         """
         Create a transport based route VPN. This VPN type uses IPSEC
         for protecting the payload, therefore a VPN Profile is specified.
@@ -210,13 +286,45 @@ class RouteVPN(Element):
             'rbvpn_tunnel_side_a': local_endpoint.data,
             'rbvpn_tunnel_side_b': remote_endpoint.data,
             'tunnel_encryption': 'transport_mode',
-            'vpn_profile_ref': profile.href
-            }
+            'vpn_profile_ref': profile.href,
+            'tunnel_mode': 'gre',
+            'enabled': enabled,
+            'comment': comment
+        }
         
         try:
             return ElementCreator(cls, json)
         except CreateElementFailed as err:
             raise CreateVPNFailed(err)
+    
+    def enable(self):
+        """
+        Enable this route based VPN
+        
+        :return: None
+        """
+        if not self.enabled:
+            self.update(enabled=True)
+    
+    def disable(self):
+        """
+        Disable this route based VPN
+        
+        :return: None
+        """
+        if self.enabled:
+            self.update(enabled=False)
+    
+    def set_preshared_key(self, new_key):
+        """
+        Set the preshared key for this VPN. A pre-shared key is only
+        present when the tunnel type is 'VPN' or the encryption mode
+        is 'transport'.
+        
+        :return: None
+        """
+        if self.data.get('preshared_key'):
+            self.update(preshared_key=new_key)
       
     @property
     def local_endpoint(self):
@@ -379,7 +487,7 @@ class TunnelEndpoint(object):
         :return TunnelInterface: interface for this endpoint
         """
         if self.tunnel_interface_ref:
-            return TunnelInterface(href=self.tunnel_interface_ref)
+            return Element.from_href(self.tunnel_interface_ref)
    
     @property
     def endpoint(self):
@@ -396,8 +504,8 @@ class TunnelEndpoint(object):
     @property
     def remote_address(self):
         """
-        Show the remote address for GRE RBVPN for Tunnel Mode or No
-        Encryption Mode configurations.
+        Show the remote IP address configured for a GRE RBVPN using Tunnel
+        or No Encryption Mode configurations.
         """
         return self.ip_address